creating a dropdown and fetching data from database into its items

+++ frontend/src/components/ProductDetails.jsx
@@ -1,23 +1,46 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 import { useProductContext } from "../hooks/useProductsContext";
 
 const ProductDetails = ({ product }) => {
   const { dispatch } = useProductContext();
 
-  const handleClick = async () => {
-    const response = await fetch("/api/products" + product._id, {
-      method: "DELETE",
+  const handleDelete = async () => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#233066",
+      cancelButtonColor: "#EC2026",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const response = await fetch("/api/products/" + product._id, {
+          method: "DELETE",
+        });
+        const json = await response.json();
+        if (response.ok) {
+          dispatch({ type: "DELETE_PRODUCT", payload: json });
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success",
+          });
+        } else {
+          Swal.fire({
+            title: "Error!",
+            text: "Failed to delete the product.",
+            icon: "error",
+          });
+        }
+      }
     });
-    const json = await response.json();
-
-    if (response.ok) {
-      dispatch({ type: "DELETE_WORKOUT", payload: json });
-    }
   };
 
   return (
     <div className="product-details">
-      {/* Table starts here */}
-
       <table className="table align-middle mb-0 bg-white">
         <thead className="bg-light">
           <tr>
@@ -31,21 +54,22 @@ const ProductDetails = ({ product }) => {
           </tr>
         </thead>
         <tbody>
-          {/* Table row starts here */}
           <tr key={product._id}>
             <td>
-              <div className="d-flex align-items-center">
-                <img
-                  src="https://mdbootstrap.com/img/new/avatars/8.jpg"
-                  alt=""
-                  style={{ width: "45px", height: "45px" }}
-                  className="rounded-circle"
-                />
-                <div className="ms-3">
-                  <p className="fw-bold mb-1">{product.name}</p>
-                  <p className="text-muted mb-0">{product.itemCode}</p>
+              <Link to={`/products/${product._id}`} className="product-link">
+                <div className="d-flex align-items-center">
+                  <img
+                    src="https://mdbootstrap.com/img/new/avatars/8.jpg"
+                    alt=""
+                    style={{ width: "45px", height: "45px" }}
+                    className="rounded-circle"
+                  />
+                  <div className="ms-3">
+                    <p className="fw-bold mb-1">{product.name}</p>
+                    <p className="text-muted mb-0">{product.itemCode}</p>
+                  </div>
                 </div>
-              </div>
+              </Link>
             </td>
             <td>
               <p className="fw-normal mb-1">{product.category}</p>
@@ -60,13 +84,14 @@ const ProductDetails = ({ product }) => {
             <td>{product.unitPrice}</td>
             <td>{product.createdAt}</td>
             <td>
-              <button onClick={handleClick}>Delete</button>
+              <button onClick={handleDelete}>Delete</button>
+              <Link to={`/UpdateProduct/${product._id}`} className="edit-link">
+                <button>Edit</button>
+              </Link>
             </td>
           </tr>
-          {/* Table row ends here */}
         </tbody>
       </table>
-      {/* Table ends here */}
     </div>
   );
 };
